+++ mods/CreateEnchantmentIndustry/CHANGELOG.md
@@ -3,9 +3,10 @@
 ### Update
 * Add Blaze Composer & Affix Templates for handling apotheosis affixes.
 * Apotheotic Content now has an independent CreativeTab.
-* Improved goggle tooltips of Blaze Forger & Enchanter.
 * Added more customizable options of Blaze Forger & Enchanter.
 * Improved Arm Interaction for Blaze Enchanter and Blaze Forger.
+* Improved goggle tooltips of Blaze Forger & Enchanter.
+* Added Goggle tooltip for Gem Cutter.
 
 If you are **modpack developer**: